pyproject.toml
[project]
name = "catap"
version = "0.1.0"
description = "Python wrapper for Apple's Core Audio Tap API"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Sage Betko", email = "sage.betko@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["audio", "macos", "core-audio", "recording", "tap", "capture"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Sound/Audio :: Capture/Recording",
]
dependencies = [
    "pyobjc-core>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-CoreAudio>=11.0; sys_platform == 'darwin'",
]

[dependency-groups]
dev = [
    "build>=1.4.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "twine>=6.2.0",
    "ty>=0.0.21",
]

[project.urls]
Homepage = "https://github.com/sbetko/catap"
Documentation = "https://github.com/sbetko/catap#readme"
Repository = "https://github.com/sbetko/catap"
Issues = "https://github.com/sbetko/catap/issues"

[project.scripts]
catap = "catap.cli:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q --strict-markers"
testpaths = ["tests"]
markers = [
    "integration: opt-in tests that exercise the real macOS Core Audio bridge",
]
